requirements.txt
streamlit
rembg
pillow
streamlit-image-comparison